+++ src/main/kotlin/io/github/erayerdin/kappdirs/appdirs/UnixAppDirs.kt
@@ -1,58 +1,48 @@
 package io.github.erayerdin.kappdirs.appdirs
 
-import io.github.erayerdin.kappdirs.AppDirs
+import java.nio.file.Path
 import java.nio.file.Paths
 
 internal class UnixAppDirs: AppDirs {
-    override fun getUserDataDir(appName: String, appVersion: String, appAuthor: String, roaming: Boolean): String {
-        val path = Paths.get(
+    override fun getUserDataDir(appName: String, appVersion: String, appAuthor: String?, roaming: Boolean): Path {
+        return Paths.get(
             HOME_DIR,
             arrayOf(".local", "share", appName, appVersion).joinToString(SEPARATOR)
         )
-
-        return path.toString()
     }
 
-    override fun getUserConfigDir(appName: String, appVersion: String, appAuthor: String, roaming: Boolean): String {
-        val path = Paths.get(
+    override fun getUserConfigDir(appName: String, appVersion: String, appAuthor: String?, roaming: Boolean): Path {
+        return Paths.get(
             HOME_DIR,
             arrayOf(".config", appName, appVersion).joinToString(SEPARATOR)
         )
-
-        return path.toString()
     }
 
-    override fun getUserCacheDir(appName: String, appVersion: String, appAuthor: String): String {
-        val path = Paths.get(
+    override fun getUserCacheDir(appName: String, appVersion: String, appAuthor: String?): Path {
+        return Paths.get(
             HOME_DIR,
             arrayOf(".cache", appName, appVersion).joinToString(SEPARATOR)
         )
-
-        return path.toString()
     }
 
-    override fun getUserLogDir(appName: String, appVersion: String, appAuthor: String): String {
-        val path = Paths.get(
+    override fun getUserLogDir(appName: String, appVersion: String, appAuthor: String?): Path {
+        return Paths.get(
             HOME_DIR,
             arrayOf(".cache", appName, "logs", appVersion).joinToString(SEPARATOR)
         )
-
-        return path.toString()
     }
 
-    override fun getSiteDataDir(appName: String, appVersion: String, appAuthor: String, local: Boolean): String {
+    override fun getSiteDataDir(appName: String, appVersion: String, appAuthor: String?, local: Boolean): Path {
         val directories: Array<String> = if (local) {
             arrayOf("usr", "local", "share", appName, appVersion)
         } else {
             arrayOf("usr", "share", appName, appVersion)
         }
 
-        val path = Paths.get("/", directories.joinToString(SEPARATOR))
-        return path.toString()
+        return Paths.get("/", directories.joinToString(SEPARATOR))
     }
 
-    override fun getSiteConfigDir(appName: String, appVersion: String, appAuthor: String, local: Boolean): String {
-        val path = Paths.get("/", arrayOf("etc", appName, appVersion).joinToString(SEPARATOR))
-        return path.toString()
+    override fun getSiteConfigDir(appName: String, appVersion: String, appAuthor: String?): Path {
+        return Paths.get("/", arrayOf("etc", appName, appVersion).joinToString(SEPARATOR))
     }
 }