+++ src/test/kotlin/io/github/erayerdin/kappdirs/appdirs/UnixAppDirsTest.kt
@@ -1,13 +1,12 @@
 package io.github.erayerdin.kappdirs.appdirs
 
-import io.github.erayerdin.kappdirs.AppDirs
 import org.junit.*
 import org.junit.Assert.assertEquals
+import java.nio.file.Paths
 
 internal val originalHome = System.getProperty("user.home")
 internal val osName = System.getProperty("os.name")?.toLowerCase()
 
-internal const val appAuthor = "eray"
 internal const val appName = "foo"
 internal const val appVersion = "0.1.0"
 
@@ -38,7 +37,7 @@ class UnixAppDirsTest {
     @Test
     fun testUserDataDir() {
         assertEquals(
-            "/home/bar/.local/share/foo/0.1.0",
+            Paths.get("/home/bar/.local/share/foo/0.1.0"),
             appDirs.getUserDataDir(appName, appVersion, appAuthor)
         )
     }
@@ -46,7 +45,7 @@ class UnixAppDirsTest {
     @Test
     fun testUserConfigDir() {
         assertEquals(
-            "/home/bar/.config/foo/0.1.0",
+            Paths.get("/home/bar/.config/foo/0.1.0"),
             appDirs.getUserConfigDir(appName, appVersion, appAuthor)
         )
     }
@@ -54,7 +53,7 @@ class UnixAppDirsTest {
     @Test
     fun testUserCacheDir() {
         assertEquals(
-            "/home/bar/.cache/foo/0.1.0",
+            Paths.get("/home/bar/.cache/foo/0.1.0"),
             appDirs.getUserCacheDir(appName, appVersion, appAuthor)
         )
     }
@@ -62,7 +61,7 @@ class UnixAppDirsTest {
     @Test
     fun testUserLogDir() {
         assertEquals(
-            "/home/bar/.cache/foo/logs/0.1.0",
+            Paths.get("/home/bar/.cache/foo/logs/0.1.0"),
             appDirs.getUserLogDir(appName, appVersion, appAuthor)
         )
     }
@@ -70,7 +69,7 @@ class UnixAppDirsTest {
     @Test
     fun testLocalSiteDataDir() {
         assertEquals(
-            "/usr/local/share/foo/0.1.0",
+            Paths.get("/usr/local/share/foo/0.1.0"),
             appDirs.getSiteDataDir(appName, appVersion, appAuthor, true)
         )
     }
@@ -78,7 +77,7 @@ class UnixAppDirsTest {
     @Test
     fun testSiteDataDir() {
         assertEquals(
-            "/usr/share/foo/0.1.0",
+            Paths.get("/usr/share/foo/0.1.0"),
             appDirs.getSiteDataDir(appName, appVersion, appAuthor)
         )
     }
@@ -86,7 +85,7 @@ class UnixAppDirsTest {
     @Test
     fun testSiteConfigDir() {
         assertEquals(
-            "/etc/foo/0.1.0",
+            Paths.get("/etc/foo/0.1.0"),
             appDirs.getSiteConfigDir(appName, appVersion, appAuthor)
         )
     }